requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
frozenlist==1.3.0
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
lxml==4.8.0
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tornado==6.1
traitlets==5.2.0
urllib3==1.26.9
wcwidth==0.2.5
yarl==1.7.2
